==> 01-load_tree_geojson.ipynb <==
import geopandas as gpd

# WFS URL
url = (
    "https://gdi.berlin.de/services/wfs/baumbestand"
    "?service=WFS&version=2.0.0&request=GetFeature"
    "&typeNames=baumbestand:strassenbaeume"
    "&outputFormat=application/json"
)

# Load into a GeoDataFrame
gdf = gpd.read_file(url)

# Inspect
print(gdf.head())
print(gdf.crs)


gdf = gdf.to_crs(epsg=4326)

gdf.to_file("data/berlin_strassenbaeume.geojson", driver="GeoJSON")
