==> 05-load_ndvi_sample.ipynb <==
# Script to try setting up Google Earth Engine Authentication and load some sample data

import ee
ee.Initialize(project='ee-ninaimmenroth')


## Load a small sample
Pick a location (e.g. Alexanderplatz) and load a small sample for a certain timeframe.

roi = ee.Geometry.Point([13.4125, 52.5218]).buffer(500)

s2 = (ee.ImageCollection("COPERNICUS/S2_SR_HARMONIZED")
      .filterBounds(roi)
      .filterDate("2023-06-01", "2023-06-30")
      .filter(ee.Filter.lt("CLOUDY_PIXEL_PERCENTAGE", 10)))

def add_ndvi(img):
    ndvi = img.normalizedDifference(["B8", "B4"]).rename("NDVI")
    return img.addBands(ndvi)

s2_ndvi = s2.map(add_ndvi)

# This extracts random sample points from the first image

img = s2_ndvi.first()

samples = img.select("NDVI").sample(
    region=roi,
    scale=10,
    numPixels=20,   # sample size
    seed=42
)


## Download as a pandas dataframe

import pandas as pd

# Convert Earth Engine FeatureCollection → client-side dict
data = samples.getInfo()

# Convert to pandas DataFrame
df = pd.DataFrame([f["properties"] for f in data["features"]])
df.head()


df.to_csv("ndvi_sample.csv", index=False)


## Download NDVI png of Berlin

# geometry
berlin = ee.Geometry.Point([13.4050, 52.5200]).buffer(5000)

# select image
s2 = (
    ee.ImageCollection("COPERNICUS/S2_SR_HARMONIZED")
    .filterBounds(berlin)
    .filterDate("2023-06-01", "2023-09-01")
    .filter(ee.Filter.lt("CLOUDY_PIXEL_PERCENTAGE", 30))
    .sort("CLOUDY_PIXEL_PERCENTAGE")
    .first()
)

# assert bands
print("bands:", s2.bandNames().getInfo())

# ndvi
ndvi = s2.normalizedDifference(["B8", "B4"]).rename("NDVI")

# style
ndvi_vis = ndvi.visualize(
    min=0.0,
    max=1.0,
    palette=['blue','white','green']
)

# thumbnail
thumb_url = ndvi_vis.getThumbURL({
    "region": berlin,
    "dimensions": 800,
    "format": "png"
})

print("Download NDVI:", thumb_url)
