==> 06b-load_ndvi_per_tree_no_watering.ipynb <==
# Script for setting up Google Earth Engine Authentication and loading the NDVI per tree
In order for this to run you must have done the steps from the notebooks load_tree_geojson and merge_datasets already to have the necessary data sets in the data folder.

import ee
ee.Initialize(project='ee-ninaimmenroth')

# Load NDVI per tree
Create a table per tree with monthly aggregate ndvi (mean of the best mosaic values in the buffer zone).
The time is limited to the year 2022 for now as it had most watering events.
I used only a sample of the trees in Charlottenburg-Wilmersdorf and Pankow since these are the districts with most waterings in 2022.

import geopandas as gpd
import matplotlib.pyplot as plt
import pandas as pd

# Load GeoJSON file of merged data set
# df_trees = gpd.read_file('data/berlin_strassenbaeume.geojson')
df_trees = gpd.read_file(
    "data/berlin_trees_with_watering_2022.gpkg",
    layer="trees_watering"
)

print(df_trees.head())

df_ch_wilm = df_trees[
    df_trees['bezirk'].isin(['Charlottenburg-Wilmersdorf', 'Pankow']) &
    ~df_trees['has_watering'] &
    df_trees['geometry'].notna() &
    ~df_trees['geometry'].is_empty
]

print(df_ch_wilm.head())
df_ch_wilm.shape
#shape is the same as without the geometry check, so no missing data there. But leaving it in just in case

## split data because of throttling
we need to use smaller samples of data, so run everything first for the first part, then for the second and merge the sets later on.

df_ch_wilm_1 = df_ch_wilm.sample(n=500, random_state=42)
df_ch_wilm_2 = (
    df_ch_wilm
    .drop(df_ch_wilm_1.index)
    .sample(n=474, random_state=42)
)

print(len(df_ch_wilm_1), len(df_ch_wilm_2))



# comment this is to do everything for the second split! and rename the file so it won't overwrite
df_ch_wilm_1 = df_ch_wilm_2

df_ch_wilm.geometry.iloc[0].__geo_interface__

features = [
    ee.Feature(
        ee.Geometry(g.__geo_interface__),
        {"tree_id": pid}
    )
    for g, pid in zip(
        df_ch_wilm_1.geometry,
        df_ch_wilm_1.pitid
    )
]

trees_fc = ee.FeatureCollection(features)

trees_fc_buffered = trees_fc.map(lambda f: f.buffer(10))

def add_ndvi(img):
    ndvi = img.normalizedDifference(["B8", "B4"]).rename("NDVI")
    return img.addBands(ndvi)

s2 = (
    ee.ImageCollection("COPERNICUS/S2_SR_HARMONIZED")
    .filterBounds(trees_fc_buffered)
    .filter(ee.Filter.lt("CLOUDY_PIXEL_PERCENTAGE", 20))
    .map(add_ndvi)
)

def monthly_ndvi_stats(year, month):
    start = ee.Date.fromYMD(year, month, 1)
    end = start.advance(1, "month")

    monthly = s2.filterDate(start, end).select("NDVI")

    reducer = (
        ee.Reducer.mean()
        .combine(ee.Reducer.median(), True)
        .combine(ee.Reducer.stdDev(), True)
        .combine(ee.Reducer.count(), True)
    )

    stats_img = monthly.reduce(reducer)

    return stats_img.sampleRegions(
        collection=trees_fc_buffered,
        scale=10,
        geometries=False
    ).map(lambda f: f.set({"year": year, "month": month}))

def monthly_ndvi_quality(year, month):
    start = ee.Date.fromYMD(year, month, 1)
    end = start.advance(1, "month")

    monthly = s2.filterDate(start, end)

    best_img = (
        monthly
        .qualityMosaic("NDVI")
        .select("NDVI")
        .rename("NDVI")
    )

    return best_img.reduceRegions(
        collection=trees_fc_buffered,
        reducer=ee.Reducer.mean().setOutputs(["NDVI"]),
        scale=10
    ).map(lambda f: f.set({
        "year": year,
        "month": month
    }))


months = list(range(3, 10))  # March–September
year = 2022

fc_list = [
    monthly_ndvi_quality(year, m)
    for m in months
]

ndvi_fc = ee.FeatureCollection(fc_list).flatten()


ndvi_fc_nogeom = ndvi_fc.map(
    lambda f: ee.Feature(
        None,
        f.toDictionary([
            "tree_id",
            "year",
            "month",
            "NDVI"
        ])
    )
)


ndvi_fc_nogeom.limit(1).getInfo()


import pandas as pd

# Convert Earth Engine FeatureCollection → client-side dict
data = ndvi_fc_nogeom.getInfo()

# Convert to pandas DataFrame
df = pd.DataFrame([f["properties"] for f in data["features"]])
df.head()
df.shape

df.to_csv("data/ndvi_CWP_2022_no_watering_p2.csv", index=False)
