==> 10-merge_all_data.ipynb <==
# Merge all the data 

import geopandas as gpd
import pandas as pd


gdf_trees = gpd.read_file("data/berlin_strassenbaeume.geojson")
gdf_trees.columns


df_ndvi = pd.read_csv("data/timeseries/ndvi_CWP_2022.csv")
df_water = pd.read_csv("data/timeseries/watering_2022.csv")  
df_weather = pd.read_csv("data/timeseries/weather_features.csv")  

df_ndvi["tree_id"] = df_ndvi["tree_id"].astype(str)
gdf_trees["pitid"] = gdf_trees["pitid"].astype(str)


# keep only trees which have ndvi
gdf_ndvi_trees = gdf_trees.merge(
    df_ndvi,
    left_on="pitid",
    right_on="tree_id",
    how="inner"
)

# merge with watering data
df_water = df_water.reset_index()

gdf_ndvi_trees = gdf_ndvi_trees.merge(
    df_water,
    left_on="pitid",
    right_on="id",
    how="left"
)


# fill missing watering
water_cols = [c for c in gdf_ndvi_trees.columns if c.startswith("water_m")]
gdf_ndvi_trees[water_cols] = gdf_ndvi_trees[water_cols].fillna(0)

# merge with weather data
gdf_ndvi_trees = gdf_ndvi_trees.merge(
    df_weather,
    how="cross"
)

gdf_ndvi_trees.head(5)

gdf_ndvi_trees.columns

# select which columns to keep
keep_cols = (
    # identifiers
    ["pitid", "geometry", "bezirk"] +

    # tree characteristics 
    # Tree taxonomy was included at the functional group, genus, and species level to control for physiological differences affecting vegetation indices.
    ["art_gruppe", "gattung", "art_bot", "pflanzjahr",
     "stammumfg", "baumhoehe", "kronedurch"] +

    # NDVI
    [f"NDVI_m{m}" for m in range(3, 10)] +

    # watering
    [f"water_m{m}" for m in range(3, 10)] +

    # weather
    [f"temp_m{m}" for m in range(3, 10)] +
    [f"precip_m{m}" for m in range(3, 10)] +

    # derived
    ["has_watering"]
)

missing = set(keep_cols) - set(gdf_ndvi_trees.columns)
assert not missing, f"Missing expected columns: {missing}"

gdf_clean = gdf_ndvi_trees[keep_cols].copy()
gdf_clean.head(5)

# save the data set
gdf_clean.to_file(
    "data/timeseries/trees_ndvi_weather_water_wide.gpkg",
    layer="trees",
    driver="GPKG"
)

df_long = pd.wide_to_long(
    gdf_clean,
    stubnames=["NDVI", "water", "temp", "precip"],
    i="pitid",
    j="month",
    sep="_m"
).reset_index()

df_long.head()

df_long.isna().mean()[["NDVI", "water", "temp", "precip"]]
df_long.groupby("month")["NDVI"].describe()

# save the data set
df_long.to_file(
    "data/timeseries/trees_ndvi_weather_water_long.gpkg",
    layer="trees",
    driver="GPKG"
)
